refactor auth logging imports
